-(trusted-)ratelimit opts now support float
-ratelimit 0.5 -> max 1req per 2sec per server
-ratelimit 0.1 -> max 1req per 10 sec per server
etc...
same for -trusted-ratelimit

+++ internal/config/settings.go
@@ -11,7 +11,7 @@ type Settings struct {
 	MaxThreads			int
 	GlobRateLimit		int
 	// per server
-	PerSrvRateLimit		int
+	PerSrvRateLimit		float64
 	PerSrvMaxFailures	int
 	// per check
 	PerCheckMaxAttempts	int
